fix:リファクタリング
